refactor(ai): move SDK reference prompts to @n8n/workflow-sdk/prompts (no-changelog)
Move EXPRESSION_REFERENCE, ADDITIONAL_FUNCTIONS, WORKFLOW_RULES, and
WORKFLOW_SDK_PATTERNS to @n8n/workflow-sdk as the single source of truth.
Both ai-workflow-builder.ee and instance-ai now import from the shared
location instead of maintaining independent forks.
