+++ packages/@n8n/workflow-sdk/src/prompts/index.ts
@@ -1,0 +1,6 @@
+export {
+	EXPRESSION_REFERENCE,
+	ADDITIONAL_FUNCTIONS,
+	WORKFLOW_RULES,
+	WORKFLOW_SDK_PATTERNS,
+} from './sdk-reference';
