+++ packages/@n8n/workflow-sdk/src/prompts/sdk-reference/index.ts
@@ -1,0 +1,4 @@
+export { EXPRESSION_REFERENCE } from './expressions';
+export { ADDITIONAL_FUNCTIONS } from './additional-functions';
+export { WORKFLOW_RULES } from './workflow-rules';
+export { WORKFLOW_SDK_PATTERNS } from './workflow-patterns';
